+++ index.js
@@ -7,6 +7,7 @@ module.exports = {
     'no-chain': require('./rules/no-chain'),
     'no-extraneous-function-wrapping': require('./rules/no-extraneous-function-wrapping'),
     'no-single-composition': require('./rules/no-single-composition'),
+    'prefer-constant': require('./rules/prefer-constant'),
     'prefer-get': require('./rules/prefer-get'),
     'use-fp': require('./rules/use-fp')
   },
@@ -25,6 +26,7 @@ module.exports = {
         'lodash-fp/no-chain': 'error',
         'lodash-fp/no-extraneous-function-wrapping': 'error',
         'lodash-fp/no-single-composition': 'error',
+        'lodash-fp/prefer-constant': ['error', false],
         'lodash-fp/prefer-get': 'error',
         'lodash-fp/use-fp': 'error'
       }