+++ rules/prefer-constant.js
@@ -1,0 +1,45 @@
+'use strict';
+
+var astUtil = require('./core/astUtil');
+
+module.exports = function (context) {
+  var shouldCheckArrowFunctions = context.options[0] === true;
+
+  function isCompletelyLiteral(node) {
+    switch (node.type) {
+      case 'Literal':
+        return true;
+      case 'BinaryExpression':
+        return isCompletelyLiteral(node.left) && isCompletelyLiteral(node.right);
+      case 'UnaryExpression':
+        return isCompletelyLiteral(node.argument);
+      case 'ConditionalExpression':
+        return isCompletelyLiteral(node.test) && isCompletelyLiteral(node.consequent) && isCompletelyLiteral(node.alternate);
+      default:
+        return false;
+    }
+  }
+
+  function handleFunctionExpression(node) {
+    var valueReturnedInFirstLine = astUtil.getValueReturnedInFirstLine(node);
+    if (valueReturnedInFirstLine && isCompletelyLiteral(valueReturnedInFirstLine)) {
+      context.report(node, 'Prefer `_.constant` over a function returning a literal');
+    }
+  }
+
+  return {
+    FunctionExpression: handleFunctionExpression,
+    FunctionDeclaration: handleFunctionExpression,
+    ArrowFunctionExpression: function (node) {
+      if (shouldCheckArrowFunctions) {
+        handleFunctionExpression(node);
+      }
+    }
+  };
+};
+
+module.exports.schema = [
+  {
+    type: 'boolean'
+  }
+];